fix(framework): resolve three functional gaps in dispatch, async lifecycle, and tool resolution
- Add protocol validation in FrameworkAdapterRegistry to enforce keyword-only params
- Fix AutoGenV4Adapter and AG2Adapter to implement full protocol signature and move logic from orphaned _run_* methods
- Remove orphaned _run_autogen_v4 and _run_ag2 methods that were never called
- Replace new_event_loop with _async_bridge in InteractiveRuntime to prevent subprocess lifecycle issues
- Replace ad-hoc run_async with documented _async_bridge import to use single source of truth
- Replace hardcoded 17-tool allow-list with YAML-requested tool resolution via ToolResolver
- Add warning logs for missing tools instead of silent drops in adapters

Fixes #1654
