+++ src/praisonai/praisonai/framework_adapters/crewai_adapter.py
@@ -71,7 +71,11 @@ class CrewAIAdapter(BaseFrameworkAdapter):
                 agent_tool_list = []
                 if tools_dict:
                     agent_tools = agent_details.get('tools', [])
-                    agent_tool_list = [tools_dict[t] for t in agent_tools if t in tools_dict]
+                    for t in agent_tools:
+                        if t in tools_dict:
+                            agent_tool_list.append(tools_dict[t])
+                        else:
+                            logger.warning("Tool %r requested by agent %r but not in tools_dict", t, agent_name)
                 
                 agent = Agent(
                     role=agent_details.get('role', agent_name),
